requirements.txt
python-telegram-bot==21.4
gspread==6.1.4
google-auth==2.34.0
uvicorn==0.30.6
starlette==0.38.5
